move mifare stuff to its folder
