+++ client/cmdhfmf.h
@@ -23,14 +23,14 @@
 #include "cmdparser.h"
 #include "common.h"
 #include "util.h"
-#include "mifare.h" 		// nonces_t struct
-#include "mfkey.h"  		// mfkey32_moebious
+#include "mifare.h" 				// nonces_t struct
+#include "mifare/mfkey.h"   		// mfkey32_moebious
 #include "cmdhfmfhard.h"
-#include "mifarehost.h"		// icesector_t,  sector_t
-#include "util_posix.h"		// msclock
-#include "mifaredefault.h"  // mifare default key array
-#include "cmdhf14a.h" 		// dropfield
-#include "cliparser/cliparser.h" // argtable
+#include "mifare/mifarehost.h"		// icesector_t,  sector_t
+#include "util_posix.h"				// msclock
+#include "mifare/mifaredefault.h"  	// mifare default key array
+#include "cmdhf14a.h" 				// dropfield
+#include "cliparser/cliparser.h" 	// argtable
 #include "hardnested/hardnested_bf_core.h"	// SetSIMDInstr
 
 extern int CmdHFMF(const char *Cmd);
