+++ include/mifare.h
@@ -13,6 +13,12 @@
 
 #include "common.h"
 
+#define MF_KEY_A	0
+#define MF_KEY_B	1
+
+#define MF_MAD1_SECTOR	0x00
+#define MF_MAD2_SECTOR	0x10
+
 //-----------------------------------------------------------------------------
 // ISO 14443A
 //-----------------------------------------------------------------------------
